user queries

+++ app/Http/Controllers/CarController.php
@@ -8,18 +8,14 @@ use Illuminate\Http\Request;
 
 class CarController extends Controller
 {
-    public function __construct(){}
+    public function __construct()
+    {
+    }
 
-    public function indexAdmin()
+    public function index()
     {
         $cars = Car::all();
         return view('admin.car.index', ['cars' => $cars]);
-    }
-
-    public function indexUser()
-    {
-        $cars = Car::all();
-        return view('user.car.index', ['cars' => $cars]);
     }
 
     public function create()
@@ -34,53 +30,55 @@ class CarController extends Controller
         $car = new Car();
         $car->brand = $data['brand'];
         $car->model = $data['model'];
-        $car->year = $data['year'];
-        $car->price = $data['price'];
+        $car->year = (int)$data['year'];
+        $car->price = floatval($data['price']);
         $car->user_id = auth()->user()->id;
         $car->save();
 
         return redirect()->route('user.car.index', ['car' => $car->id])->with('success', 'Car added successfully!');
     }
 
-    public function showAdmin(Car $car)
+    public function show(Car $car)
     {
-        return view('admin.car.detail', ['car' => $car]);
-    }
-    public function showUser(Car $car)
-    {
-        return view('user.car.detail', ['car' => $car]);
+        if (auth()->user()->role_id == 1) {
+            return view('admin.car.detail', ['car' => $car]);
+        } else if ($car->user_id == auth()->user()->id) {
+            return view('user.car.detail', ['car' => $car]);
+        }
     }
 
     public function edit(Car $car)
     {
-        return view('user.car.edit', ['car' => $car]);
+        if ($car->user_id == auth()->user()->id) {
+            return view('user.car.edit', ['car' => $car]);
+        }
     }
 
-    public function update(Car $car, Request $request)
+    public function update(Car $car, CarRequest $request)
     {
-        $data = $request->validate([
-            'brand' => 'required',
-            'model' => 'required',
-            'year' => 'required',
-            'price' => 'required',
-        ]);
+        if ($car->user_id == auth()->user()->id) {
+            $data =  $request->validated();
+            
+            $car->brand = $data['brand'];
+            $car->model = $data['model'];
+            $car->year = (int)$data['year'];
+            $car->price = floatval($data['price']);
+            $car->save();
 
-        $car->brand = $data['brand'];
-        $car->model = $data['model'];
-        $car->year = $data['year'];
-        $car->price = $data['price'];
-        $car->save();
-
-        return redirect()->route('user.car.detail', ['car' => $car->id])->with('success', 'Car updated successfully!');
+            return redirect()->route('user.car.detail', ['car' => $car->id])->with('success', 'Car updated successfully!');
+        }
     }
 
-    public function destroyCar(Car $car)
+    public function destroy(Car $car)
     {
-        $car->delete();
-        if(auth()->user()->role_id == 1) {
-            return redirect()->route('user.car.index')->with('Car deleted successfully!');
-        } else {
+        if (auth()->user()->role_id == 1) {
+            $car->delete();
             return redirect()->route('admin.car.index')->with('Car deleted successfully!');
+        } else {
+            if ($car->user_id == auth()->user()->id) {
+                $car->delete();
+                return redirect()->route('user.car.index')->with('Car deleted successfully!');
+            }
         }
     }
 }
